RIA-2059 Localisations and hard-coded text in controllers
