+++ app/handlers/error-handler.ts
@@ -1,5 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 import { INTERNAL_SERVER_ERROR, NOT_FOUND } from 'http-status-codes';
+import i18n from '../../locale/en.json';
 import Logger, { getLogLabel } from '../utils/logger';
 
 const logLabel: string = getLogLabel(__filename);
@@ -15,10 +16,10 @@ function pageNotFoundHandler(req: Request, res: Response, next: NextFunction) {
   if (list.includes('text/html')) {
     res.render('errors/404.njk');
   } else if (list.includes('application/json')) {
-    res.send({ error: NOT_FOUND, message: 'Page not found' });
+    res.send({ error: NOT_FOUND, message: i18n.error.error404.title });
   } else {
     res.type('text');
-    res.send('Page not found');
+    res.send(i18n.error.error404.title);
   }
 }
 
@@ -37,10 +38,10 @@ function serverErrorHandler(err: any, req: Request, res: Response, next: NextFun
   if (list.includes('text/html')) {
     res.render('errors/500.njk');
   } else if (list.includes('application/json')) {
-    res.send({ error: INTERNAL_SERVER_ERROR, message: 'Service unavailable' });
+    res.send({ error: INTERNAL_SERVER_ERROR, message: i18n.error.error500.title });
   } else {
     res.type('text');
-    res.send('Service unavailable');
+    res.send(i18n.error.error500.title);
   }
 }
